// unitokb/src/permissions/permission.enum.ts
/* eslint-disable prettier/prettier */
export enum Permission {
    MANAGE_USERS = 'MANAGE_USERS',
    VIEW_USERS = 'VIEW_USERS',
    CREATE_POST = 'CREATE_POST',
    EDIT_POST = 'EDIT_POST',
    DELETE_POST = 'DELETE_POST',
    VIEW_POSTS = 'VIEW_POSTS',
    CREATE_COMMENT = 'CREATE_COMMENT',
    DELETE_COMMENT = 'DELETE_COMMENT',
    LIKE_POST = 'LIKE_POST',
    ACCESS_ADMIN_DASHBOARD = 'ACCESS_ADMIN_DASHBOARD',
    CREATE_EVENT = 'CREATE_EVENT',
    MANAGE_EVENTS = 'MANAGE_EVENTS',
    VIEW_EVENTS = 'VIEW_EVENTS',
    VIEW_NOTIFICATIONS = 'VIEW_NOTIFICATIONS',
    REPORT_USER = 'REPORT_USER',
    VIEW_ANALYTICS = 'VIEW_ANALYTICS',
    SUBMIT_FEEDBACK = 'SUBMIT_FEEDBACK',
    VIEW_COMMENT = 'VIEW_COMMENT',
    CREATE_PROFILE = 'CREATE_PROFILE',
    MANAGE_PROFILE = ' MANAGE_PROFILE',
  }
  